BE 2023-01-27

See merge request s08-webmobile4-sub2/S08P12D103!3

+++ BE/pjtback/community/views.py
@@ -36,17 +36,20 @@ def board_get(request):
     return Response(serializer.data)
 
 
+
 @extend_schema(request=BoardListSerializer(), summary='게시글 생성')
 @api_view(['POST'])
+@permission_classes([IsAuthenticated])
 def board_create(request):
-    User = get_user_model()
-    user = User.objects.get(pk=request.data['userId'])
+    # User = get_user_model()
+    # user = User.objects.get(pk=request.data['userId'])
+    user = request.user
     
     serializer = BoardListSerializer(data=request.data)
     wanted_travel = get_object_or_404(Travel, pk=request.data['travel']['travelId'])
-
+    
     if serializer.is_valid(raise_exception=True):
-        serializer.save(userId=user, travel_id=wanted_travel)
+        serializer.save(userId=user, travel=wanted_travel)
         return Response(serializer.data, status=status.HTTP_201_CREATED)
 
 
@@ -69,9 +72,9 @@ def board_filtered(request):
     boards = Board.objects.annotate(day = F('travel__endDate') - F('travel__startDate'))
 
     # 일단 하드 코딩에 가깝긴 한데, 프엔에서 넘겨주는 타이밍 까지 만드는게 아니면 그냥 이런식으로 쓰고 수정하는게 나을듯?
-    age_dic = { '10대' : [10,20], '20대': [20,30] , '30대' : [30,40] , '40대': [40,50], '50대 이상': [50,100]}
-    periods_dic = {'당일치기': 1, '1박2일': 2, '3박4일': 3, '4박5일+': 4}
-    theme_lst = ['혼자여행', '커플여행', '효도여행', '우정여행', '직장여행']
+    age_dic = { '10대' : [10,20], '20대': [20,30] , '30대' : [30,40] , '40대': [40,50], '50대': [50,60], '60대 이상': [60,100]}
+    periods_dic = {'당일 치기': 1, '1박 2일': 2, '3박 4일': 3, '4박 5일+': 4}
+    theme_lst = ['혼자', '친구와', '연인과', '배우자와', '아이와', '부모님과', '기타']
     region_lst = ["서울","경기","강원","부산","경북·대구","전남·광주","제주","충남·대전","경남","충북","경남","전북","인천"]
 
 
@@ -109,20 +112,21 @@ def board_filtered(request):
     result_query = age_query & period_query & theme_query & region_query
     result_board = boards.filter(result_query)
 
-    print(result_query)
-
     serializer = BoardListSerializer(result_board, many= True)
 
     return Response(serializer.data, status= status.HTTP_200_OK)
 
-
+@extend_schema(responses=TravelSerializer(many = True), summary='게시글 전체 조회')
 @api_view(['GET'])
 def travel_get(request):
     travels = Travel.objects.all()
     serializer = TravelSerializer(travels, many=True)
     return Response(serializer.data)
 
+
+@extend_schema(request=TravelSerializer(), summary='단일 게시글 조회 수정 삭제')
 @api_view(['GET', 'PUT', 'DELETE'])
+@permission_classes([IsAuthenticated])
 def travel_detail(request, travel_id):
     travel = get_object_or_404(Travel, id = travel_id)
     if request.method == 'GET':
@@ -130,47 +134,92 @@ def travel_detail(request, travel_id):
         return Response(serializer.data)
 
     elif request.method == 'PUT':
-        # user = request.user
-        # if travel.userId == user:
-        User = get_user_model()
-        user = User.objects.get(id=1)
-        serializer = TravelSerializer(travel, data=request.data)
-        if serializer.is_valid(raise_exception=True):
-            serializer.save(userId = user)
-            
-            return Response(serializer.data, status=status.HTTP_201_CREATED)  
+        user = request.user
+        if travel.userId == user:
+        # User = get_user_model()
+        # user = User.objects.get(id=1)
+            serializer = TravelSerializer(travel, data=request.data)
+            if serializer.is_valid(raise_exception=True):
+                serializer.save(userId = user)
+                
+                return Response(serializer.data, status=status.HTTP_201_CREATED)  
     
     elif request.method == 'DELETE':
-        # if request.user == travel.userId:
-        #     travel.delete()
-        #     return Response(status=status.HTTP_204_NO_CONTENT)
-        # else:
-        #     return Response(status=status.HTTP_401_UNAUTHORIZED)
-        travel.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        if request.user == travel.userId:
+            travel.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        else:
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
+        # travel.delete()
+        # return Response(status=status.HTTP_204_NO_CONTENT)
 
+
+@extend_schema(request=TravelSerializer(), summary='여정 생성')
 @api_view(['POST'])
+@permission_classes([IsAuthenticated])
 def travel_create(request):
-    # user = request.user
-    User = get_user_model()
-    user = User.objects.get(id=1)
+    user = request.user
+    # User = get_user_model()
+    # user = User.objects.get(id=1)
     serializer = TravelSerializer(data=request.data, context ={'request': request})
     if serializer.is_valid(raise_exception=True):
         serializer.save(userId=user)
 
         return Response(serializer.data, status=status.HTTP_201_CREATED)
 
+
+@extend_schema(summary='user_id 파라미터로 넣어 주면 해당 유저의 travel')
 @api_view(['GET'])
 def travel_user(request, user_id):
     travels = Travel.objects.filter(userId__id = user_id)
     serializer = TravelSerializer(travels, many = True)
 
     return Response(serializer.data)
-        
 
+@extend_schema(summary='Board 상세페이지 조회, 수정, 삭제')
+@api_view(['GET', 'PUT', 'DELETE'])
+@permission_classes([IsAuthenticated])
+def board_detail(request, board_id):
 
+    board = get_object_or_404(Board, id = board_id)
+    if request.method == 'GET':
+        serializer = BoardListSerializer(board)
+        return Response(serializer.data)
 
+    elif request.method == 'PUT':
+        user = request.user
+        if board.userId == user:
+        # User = get_user_model()
+        # user = User.objects.get(id=1)
+            wanted_travel = get_object_or_404(Travel, pk=request.data['travel']['travelId'])
+            serializer = BoardListSerializer(board, data=request.data)
+            if serializer.is_valid(raise_exception=True):
+                serializer.save(userId=user, travel=wanted_travel)
+                
+            return Response(serializer.data, status=status.HTTP_201_CREATED)  
+    
+    elif request.method == 'DELETE':
+        if request.user == board.userId:
+            board.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        else:
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
+        # board.delete()
+        # return Response(status=status.HTTP_204_NO_CONTENT)
 
+@api_view(['POST'])
+@permission_classes([IsAuthenticated])
+def like(request, board_id):
+    board = Board.objects.get(id = board_id)
+    user = request.user
+    # User = get_user_model()
+    # user = User.objects.get(id=1)
+    if board.likeList.filter(id=request.user.id).exists():
+        board.likeList.remove(user)
+        return Response(data = {False},status=status.HTTP_202_ACCEPTED)
+    else:
+        board.likeList.add(user)
+        return Response(data = {True}, status=status.HTTP_202_ACCEPTED)
